requirements.txt
python-dotenv
openai
streamlit
google-generativeai
